requirements.txt
langchain 
langchain-openai
openai 
pydantic 
streamlit 
python-dotenv
langchain-community